An alternative Arabic translation.
The current translation is perfectly grammatical and does carry the full meaning of the original quote. However, it is too literal, akin more to a Google translation than a native one. You wouldn't come across such a sentence in a book or newspaper for instance. The one I'm proposing in the other hand is more like a real-world rendition of the quote. I hope that people with a good grasp on Arabic would also weigh in on this.