Fix gibibyte symbol
1. Avoid bit/byte confusion
IEC 60027 specifies that the symbol for binary digit should be 'bit',
and this should be used in all multiples, such as 'kbit', for kilobit.
However, the lower-case letter 'b' is widely used as well and was
recommended by the IEEE 1541 Standard (2002).

2. Also avoid "giga" confusion.
Giga according to SI is abbreviated as uppercase G and means 10^6.
Termbench uses binary units (1024^n) therefore this should be written as
Gi (gibi)